+++ include/indicator.hpp
@@ -4,92 +4,18 @@
 #include <ranges>
 
 #include <cliprogressbar/progress_indicator.hpp>
-#include <cliprogressbar/widgets/label.hpp>
-#include <cliprogressbar/widgets/bar.hpp>
-#include <cliprogressbar/formatters.hpp>
-#include <cliprogressbar/progress_plugins/ewma_rate.hpp>
-#include <cliprogressbar/progress_plugins/smoothed_eta.hpp>
-
-#include "dottorrent/file_entry.hpp"
+#include <dottorrent/file_storage.hpp>
+#include <dottorrent/file_entry.hpp>
 
 namespace rng = std::ranges;
 namespace clp = cliprogress;
 namespace dt = dottorrent;
 
-inline std::unique_ptr<cliprogress::progress_indicator>
+std::unique_ptr<cliprogress::progress_indicator>
 make_indicator(
         const dottorrent::file_storage& s,
-        const dottorrent::file_entry& e)
-{
-    auto i = std::make_unique<cliprogress::progress_indicator>(cliprogress::application::instance());
-    i->set_progress_range(0, e.file_size());
-    i->set_rate_strategy(std::make_unique<clp::ewma_rate>());
-    i->set_eta_strategy(std::make_unique<clp::smoothed_eta>());
+        const dottorrent::file_entry& e);
 
-    auto it = rng::find(s, e);
-    auto file_idx = std::distance(s.begin(), it);
-
-    // file index
-    auto index_label = std::make_shared<clp::label>(fmt::format("({}/{})", file_idx+1, s.file_count()));
-    index_label->set_name("index");
-
-    // file name
-    auto name_label = std::make_shared<clp::label>(e.path().filename().string());
-    name_label->set_name("filename");
-
-    // file size
-    auto size_label = std::make_shared<clp::label>();
-    size_label->set_name("size");
-
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                size_label->set_text(clp::format_binary_unit(data->value(), "B | "));
-            });
-
-    // bar
-    auto bar = std::make_shared<clp::bar>(
-            clp::bar_symbols { .complete_frames = std::span(clp::bar_frames::horizontal_blocks) },
-            clp::bar_style {},
-            10);
-    bar->set_name("bar");
-
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                bar->set_percentage(data->percentage());
-            });
-
-    auto percent = std::make_shared<clp::label>();
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                percent->set_text(fmt::format("{:>3.0f}%", data->percentage()));
-            });
-    percent->set_name("percent");
-
-    auto rate = std::make_shared<clp::label>();
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                rate->set_text(clp::format_binary_unit(data->rate(), "B/s"));
-            });
-    rate->set_name("rate");
-
-    index_label->set_size_policy(clp::size_policy_flag::fixed);
-    name_label->set_size_policy(clp::size_policy_flag::grow | clp::size_policy_flag::shrink);
-    percent->set_size_policy(clp::size_policy_flag::fixed);
-    bar->set_size_policy(clp::size_policy_flag::fixed);
-    size_label->set_size_policy(clp::size_policy_flag::fixed);
-    rate->set_size_policy(clp::size_policy_flag::fixed);
-
-    name_label->set_ellipsize_mode(clp::ellipsize_mode::end);
-
-    i->push_back(index_label);
-    i->push_back(name_label);
-    i->push_back(percent);
-    i->push_back(bar);
-    i->push_back(rate);
-    i->set_spacing(1);
-
-    return i;
-}
 
 inline void on_indicator_completion(std::unique_ptr<cliprogress::progress_indicator>& indicator)
 {
@@ -104,84 +30,21 @@ inline void on_indicator_completion(std::unique_ptr<cliprogress::progress_indica
 
 
 /// Indicator for v2 and hybrid torrents which does not show padding files.
-inline std::unique_ptr<cliprogress::progress_indicator>
+std::unique_ptr<cliprogress::progress_indicator>
 make_indicator_v2(
         const dottorrent::file_storage& s,
-        const dottorrent::file_entry& e)
-{
-    if (e.is_padding_file())
-        return nullptr;
+        const dottorrent::file_entry& e);
 
-    auto i = std::make_unique<cliprogress::progress_indicator>(cliprogress::application::instance());
-    i->set_progress_range(0, e.file_size());
-    i->set_rate_strategy(std::make_unique<clp::ewma_rate>());
-    i->set_eta_strategy(std::make_unique<clp::smoothed_eta>());
 
-    auto it = rng::find(s, e);
-    auto file_idx = std::distance(s.begin(), it);
 
-    std::size_t regular_file_count = s.regular_file_count();
-    const auto is_padding_file = [](const dt::file_entry& e) { return e.is_padding_file(); };
-    std::size_t regular_file_idx = file_idx - std::count_if(s.begin(), s.begin()+file_idx, is_padding_file);
+void print_simple_indicator(
+            std::ostream& os, const dottorrent::file_storage& s,
+            std::size_t file_idx, dt::protocol hasher_protocol);
 
-    // file index
-    auto index_label = std::make_shared<clp::label>(fmt::format("({}/{})", regular_file_idx+1, regular_file_count));
-    index_label->set_name("index");
+void print_simple_indicator_v1(
+        std::ostream& os, const dottorrent::file_storage& s,
+        std::size_t file_idx);
 
-    // file name
-    auto name_label = std::make_shared<clp::label>(e.path().filename().string());
-    name_label->set_name("filename");
-
-    // file size
-    auto size_label = std::make_shared<clp::label>();
-    size_label->set_name("size");
-
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                size_label->set_text(clp::format_binary_unit(data->value(), "B | "));
-            });
-
-    // bar
-    auto bar = std::make_shared<clp::bar>(
-            clp::bar_symbols { .complete_frames = std::span(clp::bar_frames::horizontal_blocks) },
-            clp::bar_style {},
-            10);
-    bar->set_name("bar");
-
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                bar->set_percentage(data->percentage());
-            });
-
-    auto percent = std::make_shared<clp::label>();
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                percent->set_text(fmt::format("{:>3.0f}%", data->percentage()));
-            });
-    percent->set_name("percent");
-
-    auto rate = std::make_shared<clp::label>();
-    i->value_changed.connect(
-            [=](const clp::progress_data* data) {
-                rate->set_text(clp::format_binary_unit(data->rate(), "B/s"));
-            });
-    rate->set_name("rate");
-
-    index_label->set_size_policy(clp::size_policy_flag::fixed);
-    name_label->set_size_policy(clp::size_policy_flag::grow | clp::size_policy_flag::shrink);
-    percent->set_size_policy(clp::size_policy_flag::fixed);
-    bar->set_size_policy(clp::size_policy_flag::fixed);
-    size_label->set_size_policy(clp::size_policy_flag::fixed);
-    rate->set_size_policy(clp::size_policy_flag::fixed);
-
-    name_label->set_ellipsize_mode(clp::ellipsize_mode::end);
-
-    i->push_back(index_label);
-    i->push_back(name_label);
-    i->push_back(percent);
-    i->push_back(bar);
-    i->push_back(rate);
-    i->set_spacing(1);
-
-    return i;
-}
+void print_simple_indicator_v2(
+        std::ostream& os, const dottorrent::file_storage& s,
+        std::size_t file_idx);